+++ src/consensus/aft/raft.h
@@ -1420,6 +1420,7 @@ namespace aft
           }
 
           case kv::ApplyResult::PASS_SNAPSHOT_EVIDENCE:
+          case kv::ApplyResult::PASS_ENCRYPTED_PAST_LEDGER_SECRET:
           {
             break;
           }